Remove egg.egg-info
