--- rimcli.egg-info/SOURCES.txt
@@ -1,29 +1,0 @@
-LICENSE
-MANIFEST.in
-README.md
-setup.cfg
-setup.py
-rimcli/RDIMCV.py
-rimcli/__init__.py
-rimcli.egg-info/PKG-INFO
-rimcli.egg-info/SOURCES.txt
-rimcli.egg-info/dependency_links.txt
-rimcli.egg-info/entry_points.txt
-rimcli.egg-info/requires.txt
-rimcli.egg-info/top_level.txt
-rimcli/modcli/CDDCV.py
-rimcli/modcli/RDDCV.py
-rimcli/modcli/UDDCV.py
-rimcli/modcli/__init__.py
-rimcli/tests/__init__.py
-rimcli/tests/e01m001.py
-rimcli/tests/data/B00001.im7
-rimcli/tests/data/B00001_uint16_0.png
-rimcli/tests/data/B00001_uint16_1.png
-rimcli/tests/data/__init__.py
-rimcli/tests/data/datlnk.py
-rimcli/tests/data/__pycache__/__init__.cpython-313.pyc
-rimcli/tests/data/__pycache__/datlnk.cpython-313.pyc
-rimcli/tests/devenv/__init__.py
-rimcli/tests/devenv/rimbase.yml
-rimcli/tests/devenv/rimenv.yml
