--- rimcli.egg-info/entry_points.txt
@@ -1,2 +1,0 @@
-[console_scripts]
-rim-cli = rimcli:rcmain
