--- rimcli.egg-info/requires.txt
@@ -1,3 +1,0 @@
-numpy<2.0
-matplotlib>=3.6
-ReadIM
